refactor(config): separate handler file by domain
