+++ src/main/java/com/dongsoop/dongsoop/common/handler/websocket/CustomStompErrorHandler.java
@@ -1,4 +1,4 @@
-package com.dongsoop.dongsoop.handler;
+package com.dongsoop.dongsoop.common.handler.websocket;
 
 import com.dongsoop.dongsoop.chat.exception.UserKickedException;
 import java.nio.charset.StandardCharsets;
