--- src/main/java/com/dongsoop/dongsoop/handler/CustomAccessDeniedHandler.java
@@ -1,31 +1,0 @@
-package com.dongsoop.dongsoop.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-@Component
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
-
-    private final HandlerExceptionResolver resolver;
-
-    public CustomAccessDeniedHandler(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver resolver) {
-        this.resolver = resolver;
-    }
-
-    /**
-     * Spring MVC 레이어의 예외 처리 체계로 위임
-     */
-    @Override
-    public void handle(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            AccessDeniedException accessDeniedException) {
-
-        resolver.resolveException(request, response, null, accessDeniedException);
-    }
-}
