--- src/main/java/com/dongsoop/dongsoop/handler/CustomAuthenticationEntryPoint.java
@@ -1,31 +1,0 @@
-package com.dongsoop.dongsoop.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.NonNull;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final HandlerExceptionResolver resolver;
-
-    public CustomAuthenticationEntryPoint(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver resolver) {
-        this.resolver = resolver;
-    }
-
-    /**
-     * Spring MVC 레이어의 예외 처리 체계로 위임
-     */
-    @Override
-    public void commence(@NonNull HttpServletRequest request,
-                         @NonNull HttpServletResponse response,
-                         @NonNull AuthenticationException authException) {
-
-        resolver.resolveException(request, response, null, authException);
-    }
-}
